streaming_service/requirements.txt
aiohttp==3.8.4
# aioredis==1.3.1
# aioredis-cluster=2.5.0
aiosignal==1.3.1
altgraph==0.17.3
anyio==3.6.2
async-timeout==4.0.2
attrs==22.2.0
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
cryptography==42.0.5
Cython==3.0.0b2
dnspython==2.3.0
email-validator==1.3.1
fastapi==0.95.0
frozenlist==1.3.3
gunicorn==20.1.0
h11==0.14.0
hiredis==2.2.2
httpcore==0.16.3
httptools==0.5.0
httpx==0.23.3
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
multidict==6.0.4
openai==1.14.2  #note - mismatch from data service version
orjson==3.8.9
pydantic==1.10.7
pyinstaller==5.13.0
pyinstaller-hooks-contrib==2023.5
PyJWT==2.8.0
python-dotenv==1.0.0
python-multipart==0.0.6
PyYAML==6.0
regex==2023.3.23
redis==5.0.1
requests==2.28.2
rfc3986==1.5.0
slowapi
sniffio==1.3.0
starlette==0.26.1
tiktoken==0.3.3
tqdm==4.65.0
typing_extensions==4.7.0
ujson==5.7.0
urllib3==1.26.15
uvicorn==0.21.1
uvloop==0.17.0
watchfiles==0.19.0
yarl==1.8.2
